requirements.txt
torch
pillow
transformers
opencv-python
tqdm
diffusers
accelerate
einops
tqdm